Harden the IPC receive paths, and probe them with hostile input

+++ shell/windows/Daemon.cs
@@ -44,7 +44,7 @@ internal static class Daemon
             Native.COPYDATASTRUCT data = new()
             {
                 dwData = IntPtr.Zero,
-                cbData = bytes.Length,
+                cbData = (uint)bytes.Length,
                 lpData = buffer,
             };
